% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_distances.R
\name{calculate_distances}
\alias{calculate_distances}
\title{Calculate Bray-Curtis dissimilarities and (un)weighted UniFrac distances}
\usage{
calculate_distances(tse, assay_relAbd)
}
\arguments{
\item{tse}{Object of class \code{TreeSummarizedExperiment}}

\item{assay_relAbd}{Character name of the assay in the \code{tse} object that
contains relative abundances, as a base for calculating the Bray-Curtis
dissimiliraties.}
}
\value{
List of two data.frames
}
\description{
Based on a \code{TreeSummarizedExperiment} object and the relative
abundances contained in it this function calculates the pairwise Bray-Curtis
dissimilarities (based on function \code{vegan::vegdist}) and the pairwise
(un)weighted UniFrac distances (based on function \code{mia::calculateUnifrac})
between all samples, and returns them in form of a list of two data frames,
comprising the intra-personal distances and the inter-personal distances.
}
