% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_participantReports.R
\name{create_participantReports}
\alias{create_participantReports}
\title{Create one measurement summary pdf report for each participant}
\usage{
create_participantReports(
  path_sauerkrautBuddyDir,
  path_targetDir,
  path_groupInfoRdata,
  path_participantRdata,
  path_samplesIDxlsx,
  path_absTaxonomyData,
  path_relTaxonomyData,
  path_sampleLookupXlsx,
  path_krautLookupXlsx,
  path_bloodMarkerBostonRdata,
  path_bloodMarkerZentrallaborRdata,
  path_bloodMetabolomeXlsx,
  path_stoolMetabolomeXlsx,
  path_bodyMeasuresXlsx
)
}
\arguments{
\item{path_sauerkrautBuddyDir}{Path to the \code{sauerkrautBuddy} directory
as downloaded from GitHub.}

\item{path_targetDir}{Path to the target directory where the individual
participant pdf reports should be stored.}

\item{path_groupInfoRdata}{Path to the \code{Gruppe_Info.Rdata} file}

\item{path_participantRdata}{Path to the \code{DATA_participant.Rdata} file}

\item{path_samplesIDxlsx}{Path to the \code{Samples_ID.xlsx} file}

\item{path_absTaxonomyData, path_relTaxonomyData}{Path to the absolute and
relative abundance taxonomy files \code{merged_abundance_table_counts.tsv}
and \code{merged_abundance_table_relab.tsv}}

\item{path_sampleLookupXlsx}{Path to the \code{Metabolon_Blutproben_Stuhlproben_intern.xlsx}
file which contains one sheet \code{Sample Manifest} summarizing all relevant
sample information, including the information which stool and blood samples were
analyzed twice in the lab for us to measure the measurement variability.}

\item{path_krautLookupXlsx}{Path to the \code{Sauerkrautproben für Sequenzierung.xlsx}
file which contains lookup information on the sauerkraut samples.}

\item{path_bloodMarkerBostonRdata}{Path to the file
\code{BloodData_Boston.Rdata}}

\item{path_bloodMarkerZentrallaborRdata}{Path to the file
\code{BloodData_Zentrallabor.Rdata}}

\item{path_bloodMetabolomeXlsx}{Path to the \code{FREI-01-24TASA CDT.XLSX} file}

\item{path_stoolMetabolomeXlsx}{Path to the \code{FREI-0302-21TASA CDT.XLSX} file}

\item{path_bodyMeasuresXlsx}{Path to the \code{Koerpermesswerte.xlsx} file}
}
\description{
This function calls the R package internal Quarto report once for each
participant and stores the rendered pdf reports in the specified directory.
}
