% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAndPrepare_experimentData.R
\name{dataPreparation_enrichExperimentData}
\alias{dataPreparation_enrichExperimentData}
\title{Helper function to further enrich the experiment data}
\usage{
dataPreparation_enrichExperimentData(dat)
}
\arguments{
\item{dat}{data.frame comprising the experiment data}
}
\value{
data.frame
}
\description{
This function basically wraps Virginies function \code{add_meta()} to
perform some data preparation steps for the participants' metadata forming
the experiment data.
Currently only called from within \code{\link{readAndPrepare_experimentData}}.
}
