% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_dataPaths.R
\name{define_dataPaths}
\alias{define_dataPaths}
\title{Helper function to define objects in the global environment stating the paths to all data files}
\usage{
define_dataPaths()
}
\description{
The paths are all set to the data repository with the relative path
'../data/', which is the path necessary for running the main code files
from the code appendix for the paper.
}
