% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_crossoverBoxplots.R
\name{plot_crossoverBoxplots}
\alias{plot_crossoverBoxplots}
\title{Boxplots with participant development lines for crossover studies}
\usage{
plot_crossoverBoxplots(
  dat,
  y_var,
  intervention_var,
  t_var,
  subject_var,
  cutOff_relevantChange = 0.01,
  y_isRelAbundance = FALSE,
  title = NULL,
  return_summaryTable = TRUE
)
}
\arguments{
\item{dat}{data.frame containing all the variables stated as argument}

\item{y_var}{Character name of the target variable}

\item{intervention_var}{Character name of the intervention group variable.
The variable in the dataset should have exactly two unique values.}

\item{t_var}{Character name of the measurement time variable. The variable
in the dataset should have exactly two unique values.}

\item{subject_var}{Character name of a subject id variable}

\item{cutOff_relevantChange}{Numeric, non-negative cut-off parameter used for
differentiating the three categories \code{c("pos. change", "no change", "neg. change")}.
Defaults to 0.01.}

\item{y_isRelAbundance}{Indicator if the \code{y_var} variable contains
relative abundances, in which case the plot labels are slightly altered.
Defaults to FALSE.}

\item{title}{Optional character title for the plot}

\item{print_summaryTable}{Indicator if a summary table of the individual
changes of the participants in \code{y_var} should be returned in a joint
list with the ggplot object. Defaults to \code{TRUE}.}
}
\value{
ggplot2 object
}
\description{
This function uses boxplots and one line between the boxplots per participant
to visualize the development of some feature in a crossover study.
Additionally, it returns a summary table of the visualized data.
}
