% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dimReduction.R
\name{plot_dimReduction}
\alias{plot_dimReduction}
\title{Plot the results of an estimated dimension reduction method}
\usage{
plot_dimReduction(
  tse,
  dimReduction_name,
  assay_relAbd,
  var_participantID = "participant_id",
  var_timepoint = "timepoint",
  var_group = "group",
  var_interTreat = "inter_treat",
  taxRank_dominance = "family",
  taxRank_topNfocus = NULL,
  covars_categorical = NULL,
  covars_metric = NULL
)
}
\arguments{
\item{tse}{Object of class \code{TreeSummarizedExperiment}}

\item{dimReduction_name}{Character name of a dimension reduction technique
result stored in the \code{tse} object. The results was usually stored in the
\code{tse} object by calling a function like \link[scater]{runMDS}.}

\item{assay_relAbd}{Character name of the assay in the \code{tse} object that
contains relative abundances.}

\item{var_participantID}{Character name of the participant id variable in
\code{colData(tse)}. Defaults to \code{"participant_id"}.}

\item{var_timepoint}{Character name of the timepoint variable in
\code{colData(tse)}, containing the indicator at which discrete timepoint
in the (crossover) study design the measuremtn was taken. Defaults to
\code{"timepoint"}.}

\item{var_group}{Character name of the group variable in \code{colData(tse)}.}

\item{var_interTreat}{Character name of the combined intervention-treament
variable in \code{colData(tse)}, containing values like
\code{"Baseline.Condition1"}, \code{"FollowUp.Condition1"}, etc.}

\item{taxRank_dominance}{Character name of the taxonomic rank for which the
dominant bacteria type should be plotted. Defaults to \code{"family"}.}

\item{taxRank_topNfocus}{Optional number specifying the number of taxa
categories that should be highlighted. The 'taxRank_topNfocus' most frequent
categories will be highlighted and all other categories will be grayed out.
Currently limited to maximum 13.}

\item{covars_categorical, covars_metric}{Optional vectors of variable names
of \code{colData(tse)} for each of which one extra plot should be created
and returned as a separate element in the output list.}
}
\value{
Named list of ggplot objects
}
\description{
This function was initially based on the function \code{plot_ordination}
written by Virginie (which has nothing to do with the equally named function
from the \code{phyloseq} package), but was substantially thinned out by
dropping some of its optional functionalities.
}
