% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAndPrepare_bloodMarkerData.R
\name{readAndPrepare_bloodMarkerData}
\alias{readAndPrepare_bloodMarkerData}
\title{Read and prepare the blood marker data}
\usage{
readAndPrepare_bloodMarkerData(
  path_bloodMarkerBostonRdata,
  path_bloodMarkerZentrallaborRdata,
  path_bloodMarkerBostonXlsx,
  path_bloodMarkerZentrallaborXlsx,
  path_sampleLookupXlsx,
  path_krautLookupXlsx,
  exclude_medicatedParticipants = TRUE,
  exclude_sickParticipants
)
}
\arguments{
\item{path_bloodMarkerBostonRdata}{Path to the file
\code{BloodData_Boston.Rdata}}

\item{path_bloodMarkerZentrallaborRdata}{Path to the file
\code{BloodData_Zentrallabor.Rdata}}

\item{path_bloodMarkerBostonXlsx}{Path to the file
\code{23-019 Michels 10 sent 111423(final serum results).xlsx}, for preparing
the data on duplicate measurements.}

\item{path_bloodMarkerZentrallaborXlsx}{Path to the file
\code{Probensammlung_mit Zentrallab_Duplikate.xlsx}, for preparing the data
on duplicate measurements.}

\item{path_sampleLookupXlsx}{Path to the \code{Metabolon_Blutproben_Stuhlproben_intern.xlsx}
file which contains one sheet \code{Sample Manifest} summarizing all relevant
sample information, including the information which stool and blood samples were
analyzed twice in the lab for us to measure the measurement variability.}

\item{path_krautLookupXlsx}{Path to the \code{Sauerkrautproben für Sequenzierung.xlsx}
file which contains lookup information on the sauerkraut samples.}

\item{exclude_medicatedParticipants}{Indicator if the two participants SK005
and SK093 that were on medication during the whole study should be removed
from the data. Defaults to TRUE.}

\item{exclude_sickParticipants}{One of \code{c("no exclusion", "stool type analysis", "taxonomy paper", "blood paper")}.
If one of the latter is specified, all measurements biased by a recent sickness
are excluded. Since the taxonomy paper (mainly stool measurements) and the blood paper
(blood measurements) analyze different biosamples, the exclusion list differs
between the two papers. For the stool type analysis, a few more measurements
are excluded compared to the taxonomy paper.

Note that medicated or sick participants' measurements' information is not
excluded from the returned blood Boston duplicate data since this is only
a secondary lookup table used for matching original with duplicate measurements
measured by the Boston laboratory, and not used for filtering other data.}
}
\description{
Since the blood marker data were already previously prepared by Virginie,
this function mainly reads Virginie's final \code{.Rdata} data files (for the
two laboratories in Boston and the 'Zentrallabor' in Germany), merges them
and performs some cosmetic preparations of the data.
Additionally, the function uses the raw data Excel files to prepare duplicate
datasets, and then returns these and the main measurements as a named list.
}
