% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAndPrepare_mainFunctionalPathwayData.R
\name{readAndPrepare_mainFunctionalPathwayData}
\alias{readAndPrepare_mainFunctionalPathwayData}
\title{Read and prepare all data of the functional pathway stool measurements}
\usage{
readAndPrepare_mainFunctionalPathwayData(
  path_KEGG_KO,
  path_KEGG_pathways,
  path_groupInfoRdata,
  path_participantRdata,
  path_samplesIDxlsx,
  path_dietInfoCsv,
  path_nutrientInfoCsv,
  path_bloodMetabolomeXlsx,
  path_stoolMetabolomeXlsx,
  path_T4questXlsx,
  path_stoolInfoXlsx,
  path_bodyMeasuresXlsx,
  path_sampleLookupXlsx,
  path_krautLookupXlsx,
  path_bloodMarkerBostonRdata,
  path_bloodMarkerZentrallaborRdata,
  path_bloodMarkerBostonXlsx,
  path_bloodMarkerZentrallaborXlsx,
  data_type = "stool",
  exclude_medicatedParticipants = TRUE,
  exclude_sickParticipants
)
}
\arguments{
\item{path_KEGG_KO}{Path to the file \code{KO-sum-wide.txt}}

\item{path_KEGG_pathways}{Path to the file \code{kegg_otu_table.txt}}

\item{path_groupInfoRdata}{Path to the \code{Gruppe_Info.Rdata} file}

\item{path_participantRdata}{Path to the \code{DATA_participant.Rdata} file}

\item{path_samplesIDxlsx}{Path to the \code{Samples_ID.xlsx} file}

\item{path_dietInfoCsv}{Path to the \code{FFQ2_nondietary_data.csv} file}

\item{path_nutrientInfoCsv}{Path to the \code{FFQ2_nutrintake.csv} file}

\item{path_bloodMetabolomeXlsx}{Path to the \code{FREI-01-24TASA CDT.XLSX} file}

\item{path_stoolMetabolomeXlsx}{Path to the \code{FREI-0302-21TASA CDT.XLSX} file}

\item{path_T4questXlsx}{Path to the \code{Dateneingabe Meinungsfragebögen Sauerkrautstudie.xlsx} file}

\item{path_stoolInfoXlsx}{Path to the \code{Probensammlung_mit Zentrallab_Duplikate.xlsx} file}

\item{path_bodyMeasuresXlsx}{Path to the \code{Koerpermesswerte.xlsx} file}

\item{path_sampleLookupXlsx}{Path to the \code{Metabolon_Blutproben_Stuhlproben_intern.xlsx}
file which contains one sheet \code{Sample Manifest} summarizing all relevant
sample information, including the information which stool and blood samples were
analyzed twice in the lab for us to measure the measurement variability.}

\item{path_krautLookupXlsx}{Path to the \code{Sauerkrautproben für Sequenzierung.xlsx}
file which contains lookup information on the sauerkraut samples.}

\item{path_bloodMarkerBostonRdata}{Path to the file
\code{BloodData_Boston.Rdata}}

\item{path_bloodMarkerZentrallaborRdata}{Path to the file
\code{BloodData_Zentrallabor.Rdata}}

\item{path_bloodMarkerBostonXlsx}{Path to the file
\code{23-019 Michels 10 sent 111423(final serum results).xlsx}, for preparing
the data on duplicate measurements.}

\item{path_bloodMarkerZentrallaborXlsx}{Path to the file
\code{Probensammlung_mit Zentrallab_Duplikate.xlsx}, for preparing the data
on duplicate measurements.}

\item{data_type}{One of \code{c("stool","kraut")}, indicating which samples
should be prepared and returned. Defaults to \code{"stool"}.}

\item{exclude_medicatedParticipants}{Indicator if the two participants SK005
and SK093 that were on medication during the whole study should be removed
from the data. Defaults to TRUE.}

\item{exclude_sickParticipants}{One of \code{c("no exclusion", "stool type analysis", "taxonomy paper", "blood paper")}.
If one of the latter is specified, all measurements biased by a recent sickness
are excluded. Since the taxonomy paper (mainly stool measurements) and the blood paper
(blood measurements) analyze different biosamples, the exclusion list differs
between the two papers. For the stool type analysis, a few more measurements
are excluded compared to the taxonomy paper.

Note that medicated or sick participants' measurements' information is not
excluded from the returned blood Boston duplicate data since this is only
a secondary lookup table used for matching original with duplicate measurements
measured by the Boston laboratory, and not used for filtering other data.}
}
\value{
\code{tse} data object
}
\description{
This function wraps the functions for pathway, blood marker and metabolome
data preparation and the main experiment data preparation and creates one
\code{tse} experiment data object for the main measurements in the experiment.
}
