% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAndPrepare_symptomDiaries.R
\name{readAndPrepare_symptomDiaries}
\alias{readAndPrepare_symptomDiaries}
\title{Read and prepare the symptom diary data}
\usage{
readAndPrepare_symptomDiaries(path)
}
\arguments{
\item{path}{Absolute or relative path to the Excel file}
}
\description{
Each subject in the study filled out a symptom diary to track potential
irregularities during the study period. Read the corresponding Excel file
containing all subjects and prepare the dataset.
}
