% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAndPrepare_taxonomy.R
\name{readAndPrepare_taxonomy}
\alias{readAndPrepare_taxonomy}
\title{Read and prepare the absolute or relative abundance taxonomy data}
\usage{
readAndPrepare_taxonomy(
  path_taxonomyData,
  path_sampleLookupXlsx,
  path_krautLookupXlsx,
  aggregation_level = "species",
  remove_unobservedBacteria = TRUE,
  exclude_rareStoolBacteria = TRUE,
  exclude_medicatedParticipants = TRUE,
  exclude_sickParticipants
)
}
\arguments{
\item{path_taxonomyData}{Path to either the \code{merged_abundance_table_counts.tsv}
or the \code{merged_abundance_table_relab.tsv} file.}

\item{path_sampleLookupXlsx}{Path to the \code{Metabolon_Blutproben_Stuhlproben_intern.xlsx}
file which contains one sheet \code{Sample Manifest} summarizing all relevant
sample information, including the information which stool and blood samples were
analyzed twice in the lab for us to measure the measurement variability.}

\item{path_krautLookupXlsx}{Path to the \code{Sauerkrautproben für Sequenzierung.xlsx}
file which contains lookup information on the sauerkraut samples.}

\item{aggregation_level}{One of \code{c("species","strain")}, indicating if
the datasets should be aggregated on species level or on bacteria strain level.
Defaults to \code{"species"}.}

\item{remove_unobservedBacteria}{Indicator if bacteria species / strains without
any appearance should be removed from the stool and sauerkraut datasets.
Defaults to TRUE.}

\item{exclude_rareStoolBacteria}{Indicator if bacteria species / strains in
the stool dataset should be excluded if they don't have a minimum relative
abundance of 0.01\% in at least 10\% of all samples. Defaults to TRUE.}

\item{exclude_medicatedParticipants}{Indicator if the two participants SK005
and SK093 that were on medication during the whole study should be removed
from the data. Defaults to TRUE.}

\item{exclude_sickParticipants}{One of \code{c("no exclusion", "stool type analysis", "taxonomy paper", "blood paper")}.
If one of the latter is specified, all measurements biased by a recent sickness
are excluded. Since the taxonomy paper (mainly stool measurements) and the blood paper
(blood measurements) analyze different biosamples, the exclusion list differs
between the two papers. For the stool type analysis, a few more measurements
are excluded compared to the taxonomy paper.

Note that medicated or sick participants' measurements' information is not
excluded from the returned blood Boston duplicate data since this is only
a secondary lookup table used for matching original with duplicate measurements
measured by the Boston laboratory, and not used for filtering other data.}
}
\value{
List of five data.frames
}
\description{
Taxonomy data were collected for two types of data: the stool samples of the
individual participants, and some sauerkraut samples. This function prepares
tables for the absolute or relative abundances of both. The function returns
a list of five tables, two for stool samples (original samples and some
duplicate samples to check measurement variability), one for the sauerkraut
samples and two lookup tables (one each for the stool and sauerkraut samples).
}
