% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tse_addTaxaFeatures.R
\name{tse_addTaxaFeatures}
\alias{tse_addTaxaFeatures}
\title{Add different taxa features to a tse object}
\usage{
tse_addTaxaFeatures(tse, assay_relAbd)
}
\arguments{
\item{tse}{Object of class \code{TreeSummarizedExperiment}}

\item{assay_relAbd}{Character name of the assay in the \code{tse} object that
contains relative abundances.}
}
\value{
\code{tse} object with additional columns in \code{colData(tse)}
}
\description{
This function calculates the following features based on an object \code{tse}
with class \code{TreeSummarizedExperiment} and adds them as additional
columns in \code{colData(tse)}:
\enumerate{
\item relative abundance and rel. abundance change (after the intervention) of
L. paracasei
\item dominant (= most abundant) species, genus, etc.
\item relative abundance of the dominant species, genus, etc.
\item species, genus, etc. with the biggest growth and loss
}
}
